Integrate frontend and backend

+++ src/main/java/com/justcountit/user/ForDebtorsMetadata.java
@@ -1,4 +1,4 @@
 package com.justcountit.user;
 
-public record ForDebtorsMetadata(int id, Long debtor, Double price) {
+public record ForDebtorsMetadata(Long id, Long debtor, Double price) {
 }
